completes splitting out functions into separate files
