+++ include/TouchControl.h
@@ -1,0 +1,13 @@
+#ifndef TOUCHCONTROL_H
+#define TOUCHCONTROL_H
+
+#include <Arduino.h>
+#include "VariableDeclarations.h"
+
+int arrowsTouch(TSPoint &point, bool stepMotor, int val);
+void autoScreenTouch(TSPoint &point);
+void manualScreenTouch(TSPoint &point);
+void startScreenTouch(TSPoint &point);
+void touchScreen();
+
+#endif
