+++ src/main.cpp
@@ -1,7 +1,7 @@
 // ********* Macro Stepper v1.1 *********
 // *** AUTHOR: Cam Dey
 // *** DATE: 2019-01-27
-// *** UPDATED: 2019-06-24
+// *** UPDATED: 2019-06-28
 // **************************************
 
 // Macro Stepper was written to automate the focus stacking procedure commonly used
@@ -27,33 +27,15 @@
 #include <AccelStepper.h>											// software stepping implementation
 #include <TMC2130Stepper.h> 									// stepper driver library
 #include <TMC2130Stepper_REGDEFS.h>	  				// stepper driver registry definitions
-// custom functions for this project
-#include "VariableDeclarations.h"
-#include "DriverConfig.h"
-#include "ShutterControl.h"
-// custom fonts for TFT displa]y
-// #include <Fonts/Arimo_Regular_10.h>
-// #include <Fonts/Arimo_Regular_16.h>
-// #include <Fonts/Arimo_Regular_18.h>
-// #include <Fonts/Arimo_Regular_20.h>
-// #include <Fonts/Arimo_Regular_22.h>
-// #include <Fonts/Arimo_Regular_24.h>
-// #include <Fonts/Arimo_Regular_30.h>
-// #include <Fonts/Arimo_Bold_20.h>
-// #include <Fonts/Arimo_Bold_24.h>
-// #include <Fonts/Arimo_Bold_30.h>
-// #include <Fonts/Syncopate_Bold_36.h>
-// #include <Fonts/Permanent_Marker_Regular_36.h>
-// #include <Fonts/Lato_Black_34.h>
-// // custom icons for TFT display
-// #include <Icons/Aperture_Icon.h>	  					// icon for shutter state
-// #include <Icons/Back_Arrow_Icon.h>						// icon for navigation back to previous screen
-// #include <Icons/Cog_Wheel_Icon.h>							// icon for AutoStack config
-// #include <Icons/Delay_Clock_Icon.h>						// icon for setting shutter delay
-// #include <Icons/House_Icon.h>									// icon for HomeRail on main screen
-// #include <Icons/Main_Logo_Icon.h>							// icon for main logo on home screen
-// #include <Icons/Reset_Icon.h>									// icon for reseting/clearing data
-// #include <Icons/Reset_40_Icon.h>							// icon for reseting/clearing data
+// project definitions and functions
+#include "VariableDeclarations.h"							// external variable declarations
+#include "DriverConfig.h"											// functions for configuring TMC2130 profiles
+#include "MiscFunctions.h"										// miscellaneous functions
+#include "ShutterControl.h"										// functions relating to the camera shutter and flash
+#include "StepperControl.h"										// functions for controlling the stepper motor
+#include "TouchControl.h"											// touch screen functions for detecting touches
+#include "UserInterface.h"										// generates the different screen menus
+
 
 TouchScreen 		ts 					= TouchScreen(XP, YP, XM, YM, 300);
 TMC2130Stepper 	driver 			= TMC2130Stepper(EN_PIN, DIR_PIN, STEP_PIN, CS_PIN);
@@ -277,1097 +259,5 @@ void loop() {
     bootFlag = false;
     // homeRail();
 		setAutoStackPositions(1, 1);
-  }
-}
-
-void touchScreen() {
-  currentTime = millis(); // call frequently just in case
-
-  TSPoint point = ts.getPoint();
-  pinMode(XM, OUTPUT);
-  pinMode(YP, OUTPUT);
-
-  if (point.z > minPressure && point.z < maxPressure) {
-    switch (activeScreen) {
-      case 1:
-        startScreenTouch(point); // main screen
-        break;
-      case 2:
-        manualScreenTouch(point); // screen for manual mode
-        break;
-      case 3:
-        autoScreenTouch(point); // screen for autoStack
-        break;
-      case 4:
-        autoConfigScreenTouch(point); // autoStack config
-        break;
-    }
-  }
-}
-
-void startScreen() {
-  activeScreen = 1;
-  tft.fillScreen(BLACK);
-
-  tft.drawBitmap(40, 0, logo, 240, 82, WHITE);
-
-  tft.setTextColor(BLACK);
-  // Manual option
-  tft.fillRoundRect(20, 100, 130, 60, 8, CUSTOM_BLUE);
-  tft.setCursor(30, 140);
-  tft.setFont(&Arimo_Bold_30);
-  tft.println("Manual");
-
-  // Auto option
-  tft.fillRoundRect(170, 100, 130, 60, 8, CUSTOM_GREEN);
-  tft.setCursor(200,140);
-  tft.setFont(&Arimo_Bold_30);
-  tft.println("Auto");
-
-  // Homing option
-  // tft.fillRect(0, 200, 320, 40, CUSTOM_BLUE);
-  tft.drawBitmap(30, 190, house, 50, 42, WHITE);
-
-  // Clear autoStack
-  if (autoStackFlag == false) {
-    tft.drawBitmap(250, 190, reset40, 40, 40, GRAY);
-  } else if (autoStackFlag == true) {
-    tft.drawBitmap(250, 190, reset40, 40, 40, CUSTOM_RED);
-  }
-}
-
-void startScreenTouch(TSPoint &point) {
-  // scale from 0->1023 to tft dimension and swap coordinates
-  int xPos = map(point.y, TS_MINY, TS_MAXY, 0, tft.width());
-  int yPos = map(point.x, TS_MINX, TS_MAXX, 0, tft.height());
-
-  // manual box
-  if ((xPos >= 20 && xPos <= 150) && (yPos >= 90 && yPos <= 160)) {
-    manualScreen();
-  }
-  // auto box
-  if ((xPos >= 170 && xPos <= 300) && (yPos >= 90 && yPos <= 160)) {
-    autoScreen();
-  }
-  // homing func
-  if ((xPos >= 20 && xPos <= 100) && (yPos >= 190 && yPos <= 240)) {
-    tft.drawBitmap(30, 190, house, 50, 42, CUSTOM_RED);
-    homeRail();
-    tft.drawBitmap(30, 190, house, 50, 42, WHITE);
-  }
-  // clear autoStack run
-  if ((xPos >= 230 && xPos <= 320) && (yPos >= 200 && yPos <= 240) && autoStackFlag == true) {
-    resetAutoStack(); // reset autostack settings
-  }
-}
-
-void manualScreen() {
-  activeScreen = 2;
-  tft.fillScreen(BLACK);
-  tft.setTextColor(WHITE);
-
-  // Step Distance
-  // tft.fillRoundRect(5, 5, 125, 65, 14, CUSTOM_GREY);
-  tft.fillRoundRect(-10, 5, 135, 65, 15, CUSTOM_BLUE);
-  tft.setCursor(5, 30);
-  tft.setFont(&Arimo_Regular_24);
-  tft.println("Step Dist.");
-
-  tft.setCursor(10, 60);
-  tft.setFont(&Arimo_Bold_30);
-  tft.println(String(distancePerMovement/1000, 4));
-
-  // Step Number
-  // tft.fillRoundRect(5, 85, 125, 65, 14, CUSTOM_GREY);
-  tft.fillRoundRect(-10, 85, 135, 65, 15, CUSTOM_BLUE);
-  tft.setCursor(5, 110);
-  tft.setFont(&Arimo_Regular_24);
-  tft.println("Step Nr.");
-
-  tft.setCursor(45, 140);
-  tft.setFont(&Arimo_Bold_30);
-  tft.println(manualMovementCount);
-
-  // Rail Posoition
-  // tft.fillRoundRect(5, 165, 125, 65, 14, CUSTOM_GREY);
-  tft.fillRoundRect(-10, 165, 135, 65, 15, CUSTOM_BLUE);
-  tft.setCursor(5, 190);
-  tft.setFont(&Arimo_Regular_24);
-  tft.println("Rail Pos.");
-  tft.setCursor(9, 220);
-  tft.setFont(&Arimo_Bold_30);
-  tft.println(String(stepper.currentPosition()*0.0025, 4));
-
-  // auto shutter
-  if (shutterState == true) {
-    tft.drawBitmap(155, 15, aperture, 50, 50, CUSTOM_GREEN);
-  } else if (shutterState == false) {
-    tft.drawBitmap(155, 15, aperture, 50, 50, CUSTOM_RED);
-  }
-
-  // reset steps
-  tft.drawBitmap(155, 95, reset, 50, 50, WHITE);
-
-  // back to main screen
-  tft.drawBitmap(155, 175, backArrow, 50, 50, WHITE);
-
-  drawArrows();
-}
-
-void manualScreenTouch(TSPoint &point) {
-  // scale from 0->1023 to tft dimension and swap coordinates
-  int xPos = map(point.y, TS_MINY, TS_MAXY, 0, tft.width());
-  int yPos = map(point.x, TS_MINX, TS_MAXX, 0, tft.height());
-
-  // toggle shutter on/off
-  if ((xPos >= 150 && xPos <= 220) && (yPos >= 10 && yPos <= 70)) {
-    if ((currentTime - prevGenericTime) >= genericTouchDelay) {
-      toggleShutter();
-      if (shutterState == true) {
-        tft.drawBitmap(155, 15, aperture, 50, 50, CUSTOM_GREEN);
-      } else if (shutterState == false) {
-        tft.drawBitmap(155, 15, aperture, 50, 50, CUSTOM_RED);
-      }
-
-      prevGenericTime = millis();
-    }
-  }
-  // reset manual movement count
-  if ((xPos >= 145 && xPos <= 220) && (yPos >= 90 && yPos <= 150)) {
-    int16_t x1, y1;
-    uint16_t w, h;
-    tft.setFont(&Arimo_Bold_30);
-    tft.getTextBounds(String(manualMovementCount), 45, 140, &x1, &y1, &w, &h);
-    tft.fillRect(x1, y1, w, h, CUSTOM_BLUE);
-
-    manualMovementCount = 0;
-    prevManualMovementCount = 0;
-    tft.setTextColor(WHITE);
-    tft.setCursor(45, 140);
-    tft.println(manualMovementCount);
-  }
-  // step distance button
-  if ((xPos >= 5 && xPos <= 130) && (yPos >= 5 && yPos <= 70)) {
-    if ((currentTime - prevGenericTime) >= genericTouchDelay) {
-      arrowsActive = !arrowsActive;
-      editMovementDistance = !editMovementDistance;
-
-      if (arrowsActive == false && editMovementDistance == false) {
-        tft.setTextColor(WHITE);
-        tft.setCursor(5, 30);
-        tft.setFont(&Arimo_Regular_24);
-        tft.println("Step Dist.");
-        tft.setCursor(10, 60);
-        tft.setFont(&Arimo_Bold_30);
-        tft.println(String(distancePerMovement/1000, 4));
-      }
-      if (arrowsActive == true && editMovementDistance == true) {
-        tft.setTextColor(YELLOW);
-        tft.setCursor(5, 30);
-        tft.setFont(&Arimo_Regular_24);
-        tft.println("Step Dist.");
-        tft.setCursor(10, 60);
-        tft.setFont(&Arimo_Bold_30);
-        tft.println(String(distancePerMovement/1000, 4));
-      }
-
-      prevGenericTime = millis();
-    }
-  }
-  // set step size
-  if (arrowsActive == true && editMovementDistance == true) {
-    if (prevDistance != distancePerMovement) {
-      prevDistance = distancePerMovement;
-    }
-
-    stepsPerMovement = arrowsTouch(point, 0, stepsPerMovement);
-    setStepDistance();
-  }
-  // step motor by specified distance
-  if (arrowsActive == false) {
-    int val = arrowsTouch(point, 1, 0);
-  }
-  // back button
-  if ((xPos >= 150 && xPos <= 210) && (yPos >= 170 && yPos <= 225) && arrowsActive == false) {
-    startScreen();
-  }
-}
-
-void autoScreen() {
-  activeScreen = 3;
-  tft.fillScreen(BLACK);
-  tft.setTextColor(WHITE);
-
-  // Step Distance
-  tft.fillRoundRect(-10, 5, 135, 65, 15, CUSTOM_BLUE);
-  tft.setCursor(5, 30);
-  tft.setFont(&Arimo_Regular_24);
-  tft.println("Step Dist.");
-  tft.setCursor(10, 60);
-  tft.setFont(&Arimo_Bold_30);
-  tft.println(String(distancePerMovement/1000, 4));
-
-  // Estimated time remaining
-  tft.fillRoundRect(-10, 85, 135, 65, 15, CUSTOM_BLUE);
-  tft.setCursor(5, 110);
-  tft.setFont(&Arimo_Regular_24);
-  tft.println("Est. Time");
-  // show estimated time
-  estimateDuration(1);
-
-  // Step progress
-  tft.fillRoundRect(-10, 165, 135, 65, 15, CUSTOM_BLUE);
-  tft.setCursor(5, 190);
-  tft.setFont(&Arimo_Regular_24);
-  tft.println("Progress");
-  // show progress, force update due to screen change
-  updateProgress(1);
-
-  // auto shutter
-  if (shutterState == true) {
-    tft.drawBitmap(155, 15, aperture, 50, 50, CUSTOM_GREEN);
-  } else if (shutterState == false) {
-    tft.drawBitmap(155, 15, aperture, 50, 50, CUSTOM_RED);
-  }
-
-  // enter auto config
-  tft.drawBitmap(155, 95, cogWheel, 50, 50, GRAY);
-  // back to main screen
-  tft.drawBitmap(155, 175, backArrow, 50, 50, WHITE);
-
-  if (pauseAutoStack == true && autoStackFlag == true) {
-    drawPlayPause(0, 1);
-  } else if (pauseAutoStack == false && autoStackFlag == true) {
-    drawPlayPause(1, 0);
-  } else if (pauseAutoStack == false && autoStackFlag == false) {
-    drawPlayPause(0, 0);
-  }
-}
-
-void autoScreenTouch(TSPoint &point) {
-  // scale from 0->1023 to tft dimension and swap coordinates
-  int xPos = map(point.y, TS_MINY, TS_MAXY, 0, tft.width());
-  int yPos = map(point.x, TS_MINX, TS_MAXX, 0, tft.height());
-
-  // shutter toggle
-  if ((xPos >= 140 && xPos <= 210) && (yPos >= 10 && yPos <= 70)) {
-    if ((currentTime - prevGenericTime) >= genericTouchDelay) {
-      toggleShutter();
-      if (shutterState == true) {
-        tft.drawBitmap(155, 15, aperture, 50, 50, CUSTOM_GREEN);
-      } else if (shutterState == false) {
-        tft.drawBitmap(155, 15, aperture, 50, 50, CUSTOM_RED);
-      }
-      prevGenericTime = millis();
-    }
-  }
-  // config button
-  if ((xPos >= 140 && xPos <= 210) && (yPos >= 80 && yPos <= 150)) {
-    autoConfigScreen();
-  }
-  // step distance button
-  if ((xPos >= 5 && xPos <= 130) && (yPos >= 5 && yPos <= 70)) {
-    if ((currentTime - prevGenericTime) >= genericTouchDelay) {
-      arrowsActive = !arrowsActive;
-      editMovementDistance = !editMovementDistance;
-
-      if (arrowsActive == false && editMovementDistance == false) {
-        tft.setTextColor(WHITE);
-        tft.setCursor(5, 30);
-        tft.setFont(&Arimo_Regular_24);
-        tft.println("Step Dist.");
-        tft.setCursor(10, 60);
-        tft.setFont(&Arimo_Bold_30);
-        tft.println(String(distancePerMovement/1000, 4));
-      }
-      if (arrowsActive == true && editMovementDistance == true) {
-        tft.setTextColor(YELLOW);
-        tft.setCursor(5, 30);
-        tft.setFont(&Arimo_Regular_24);
-        tft.println("Step Dist.");
-        tft.setCursor(10, 60);
-        tft.setFont(&Arimo_Bold_30);
-        tft.println(String(distancePerMovement/1000, 4));
-      }
-
-      prevGenericTime = millis();
-    }
-  }
-  // set step size
-  if (arrowsActive == true && editMovementDistance == true) {
-    if (prevDistance != distancePerMovement) {
-      prevDistance = distancePerMovement;
-    }
-
-    stepsPerMovement = arrowsTouch(point, 0, stepsPerMovement);
-    setStepDistance();
-  }
-  // run AutoStack sequence
-  if (arrowsActive == false && (xPos >= 230 && xPos <= 310) && (yPos >= 30 && yPos <= 120)) {
-    autoStackFlag = true;
-    pauseAutoStack = false; // re-enable autoStack + pause button
-    drawPlayPause(1, 0); // grey out play, recolour pause button
-  }
-  // pause/resume AutoStack sequence
-  if (arrowsActive == false && pauseAutoStack == false && autoStackFlag == true && (xPos >= 230 && xPos <= 310) && (yPos >= 135 && yPos <= 225)) {
-    pauseAutoStack = true;
-    drawPlayPause(0, 1); // recolour play, grey out pause button
-  }
-  // back button
-  if ((xPos >= 150 && xPos <= 210) && (yPos >= 170 && yPos <= 225) && arrowsActive == false) {
-    startScreen();
-  }
-}
-
-void autoConfigScreen() {
-  activeScreen = 4;
-  tft.fillScreen(BLACK);
-  tft.setTextColor(BLACK);
-
-  // set start point
-  tft.setCursor(20, 50);
-  tft.setFont(&Lato_Black_34);
-  tft.setTextColor(CUSTOM_GREEN);
-  tft.println("START");
-  tft.setCursor(35, 65);
-  tft.setFont(&Arimo_Regular_16);
-  tft.setTextColor(WHITE);
-  tft.println(startPosition);
-
-  // Set end point
-  tft.setCursor(20, 130);
-  tft.setFont(&Lato_Black_34);
-  tft.setTextColor(CUSTOM_RED);
-  tft.println("END");
-  tft.setCursor(35, 145);
-  tft.setFont(&Arimo_Regular_16);
-  tft.setTextColor(WHITE);
-  tft.println(endPosition);
-
-  // Run from start to end and return
-  // move back from end point to start point (with accel)
-  // then run from start to end with no accel
-  tft.setCursor(20, 210);
-  tft.setFont(&Lato_Black_34);
-  tft.setTextColor(CUSTOM_BLUE);
-  tft.println("RUN");
-  tft.setCursor(35, 225);
-  tft.setFont(&Arimo_Regular_16);
-  tft.setTextColor(WHITE, BLACK);
-  tft.println(stepper.currentPosition());
-
-  // set delay
-  tft.drawBitmap(160, 70, delayClock, 50, 50, WHITE);
-  tft.setFont(&Arimo_Regular_20);
-  tft.setTextColor(WHITE, BLACK);
-  tft.setCursor(155, 143);
-  tft.println(shutterDelay);
-  tft.setCursor(175, 143);
-  tft.println(" sec");
-
-  // back to auto screen
-  tft.drawBitmap(155, 175, backArrow, 50, 50, WHITE);
-
-  drawArrows();
-}
-
-void autoConfigScreenTouch(TSPoint &point) {
-  // scale from 0->1023 to tft dimension and swap coordinates
-  int xPos = map(point.y, TS_MINY, TS_MAXY, 0, tft.width());
-  int yPos = map(point.x, TS_MINX, TS_MAXX, 0, tft.height());
-
-  // start button
-  if ((xPos >= 10 && xPos <= 140) && (yPos >= 10 && yPos <= 80) && (editShutterDelay + editEndPosition) == 0) {
-    if ((currentTime - prevGenericTime) >= genericTouchDelay) {
-      arrowsActive = !arrowsActive;
-      editStartPosition = !editStartPosition;
-
-      if (arrowsActive == false && editStartPosition == false) {
-        // tft.drawRoundRect(15, 15, 125, 60, 4, BLACK);
-        tft.fillRoundRect(10, 24, 5, 40, 4, BLACK);
-      }
-      if (arrowsActive == true && editStartPosition == true) {
-        // tft.drawRoundRect(15, 15, 125, 60, 4, YELLOW);
-        tft.fillRoundRect(10, 24, 5, 40, 4, YELLOW);
-      }
-
-      prevGenericTime = millis();
-    }
-  }
-  // set start position
-  if (editStartPosition == true && arrowsActive == true) {
-    if (prevStartPosition != startPosition) {
-      prevStartPosition = startPosition;
-    }
-    int val = arrowsTouch(point, 1, 0);
-    setAutoStackPositions(1, 0); //set start but not end position
-  }
-  // end button
-  if ((xPos >= 10 && xPos <= 120) && (yPos >= 90 && yPos <= 160) && (editShutterDelay + editStartPosition) == 0) {
-    if ((currentTime - prevGenericTime) >= genericTouchDelay) {
-      arrowsActive = !arrowsActive;
-      editEndPosition = !editEndPosition;
-
-      if (arrowsActive == false && editEndPosition == false) {
-        // tft.drawRoundRect(15, 15, 125, 60, 4, BLACK);
-        tft.fillRoundRect(10, 104, 5, 40, 4, BLACK);
-      }
-      if (arrowsActive == true && editEndPosition == true) {
-        // tft.drawRoundRect(15, 15, 125, 60, 4, YELLOW);
-        tft.fillRoundRect(10, 104, 5, 40, 4, YELLOW);
-      }
-
-      prevGenericTime = millis();
-    }
-  }
-  // set end position
-  if (editEndPosition == true && arrowsActive == true) {
-    if (prevEndPosition != endPosition) {
-      prevEndPosition = endPosition;
-    }
-
-    int val = arrowsTouch(point, 1, 0);
-    setAutoStackPositions(0, 1); //set end but not start position
-  }
-  // run button
-  if ((xPos >= 10 && xPos <= 120) && (yPos >= 170 && yPos <= 240) && (editShutterDelay + editStartPosition + editEndPosition) == 0) {
-    tft.fillRoundRect(10, 184, 5, 35, 4, YELLOW);
-    dryRun();
-    tft.fillRoundRect(10, 184, 5, 35, 4, BLACK);
-  }
-  // delay button
-  if ((xPos >= 150 && xPos <= 220) && (yPos >= 60 && yPos <= 130) && (editStartPosition + editEndPosition) == 0) {
-    if ((currentTime - prevGenericTime) >= genericTouchDelay) {
-      arrowsActive = !arrowsActive;
-      editShutterDelay = !editShutterDelay;
-
-      if (arrowsActive == false && editShutterDelay == false) {
-        tft.drawBitmap(160, 70, delayClock, 50, 50, WHITE);
-      }
-      if (arrowsActive == true && editShutterDelay == true) {
-        tft.drawBitmap(160, 70, delayClock, 50, 50, YELLOW);
-      }
-
-      prevGenericTime = millis();
-    }
-  }
-  // set shutter delay
-  if (editShutterDelay == true && arrowsActive == true) {
-    if (prevDelay != shutterDelay) {
-      prevDelay = shutterDelay;
-    }
-    shutterDelay = arrowsTouch(point, 0, shutterDelay);
-    setShutterDelay();
-  }
-  // show current position of stepper
-  displayPosition();
-  // back button - tft.setCursor(150, 205);
-  if ((xPos >= 150 && xPos <= 210) && (yPos >= 170 && yPos <= 225) && arrowsActive == false) {
-    autoScreen();
-  }
-
-}
-
-void drawArrows() {
-  tft.fillTriangle(230, 65, 260, 15, 290, 65, CUSTOM_GREEN);
-  tft.fillRoundRect(245, 69, 30, 36, 4, CUSTOM_GREEN);
-
-  tft.fillTriangle(230, 175, 260, 225, 290, 175, CUSTOM_RED);
-  tft.fillRoundRect(245, 135, 30, 36, 4, CUSTOM_RED);
-}
-
-void drawPlayPause(bool greyPlay = 0, bool greyPause = 0) {
-  if (greyPlay == 1) {
-    tft.fillCircle(270, 75, 37, CUSTOM_GREY_LITE);
-    tft.fillTriangle(255, 50, 295, 75, 255, 100, WHITE);
-  } else {
-    tft.fillCircle(270, 75, 37, CUSTOM_GREEN);
-    tft.fillTriangle(255, 50, 295, 75, 255, 100, WHITE);
-  }
-
-  if (greyPause == 1) {
-    tft.fillCircle(270, 180, 37, CUSTOM_GREY_LITE);
-    tft.fillRoundRect(250, 155, 15, 50, 4, WHITE);
-    tft.fillRoundRect(275, 155, 15, 50, 4, WHITE);
-  } else {
-    tft.fillCircle(270, 180, 37, CUSTOM_BLUE);
-    tft.fillRoundRect(250, 155, 15, 50, 4, WHITE);
-    tft.fillRoundRect(275, 155, 15, 50, 4, WHITE);
-  }
-}
-
-int arrowsTouch(TSPoint &point, bool stepMotor, int val = 0) {
-  // scale from 0->1023 to tft dimension and swap coordinates
-  int xPos = map(point.y, TS_MINY, TS_MAXY, 0, tft.width());
-  int yPos = map(point.x, TS_MINX, TS_MAXX, 0, tft.height());
-
-  if ((xPos >= 230 && xPos <= 290) && (yPos >= 15 && yPos <= 105)) {
-    val++;
-    if (stepMotor == 1) { // if stepping required, trigger shutter, move step
-      if (shutterState == true && activeScreen == 2) {
-        shutterTriggered = triggerShutter();
-      }
-      stepperMoved = stepperStep(1, 500); // forward
-    }
-  }
-  if ((xPos >= 230 && xPos <= 290) && (yPos >= 135 && yPos <= 225)) {
-    val--;
-    if (stepMotor == 1) { // if stepping required, move motor
-      stepperMoved = stepperStep(-1, 500); // reverse
-    }
-  }
-
-  return val;
-}
-
-void autoStack() {
-  // wake stepper from sleep
-  if (digitalRead(EN_PIN) == HIGH) {
-    toggleStepper(1);
-  }
-  // move stepper to start if not already there
-  if (goToStart == true) {
-    while (stepper.currentPosition() != startPosition) {
-      // move to start
-      stepper.setSpeed(-600);
-      stepper.moveTo(startPosition);
-      stepper.runSpeedToPosition();
-    }
-		goToStart = false;
-		movementProgress = false;
-		joystickState = false;
-		stepperMoved = false;
-		shutterTriggered = false;
-    // stepper.move(0);
-    // stepper.setSpeed(0);
-  }
-  // take photo and increment progress for first step of each movement
-  if (movementProgress <= numMovements && stepperMoved == false) {
-		// if shutter enabled, take photo
-		if (shutterState == true && shutterTriggered == false) {
-	    shutterTriggered = triggerShutter(); // take photo
-		}
-
-		// move stepper
-		stepperMoved = stepperStep(1, shutterDelay*1000); // forward direction, shutterdelay
-
-		if (stepperMoved == true) {
-			movementProgress++;
-	    // make sure correct screen is displaying
-	    if (activeScreen == 3) { // auto screen
-	      updateProgress(0); // don't force refresh
-				estimateDuration(0); // don't force refresh
-	    }
-			shutterTriggered = false; // reset shutter
-			stepperMoved = false; // reset stepper move
-		}
-  }
-  // stop AutoStack sequence if end reached
-  if (movementProgress >= numMovements) {
-		estimateDuration(1); // force refresh
-    autoStackFlag = false;
-    goToStart = true;
-    movementProgress = false;
-    joystickState = true;
-		stepperMoved = false;
-		shutterTriggered = false;
-    drawPlayPause(0, 0); // reset play to green
-  }
-}
-
-void homeRail() {
-	// set configuration for stallGuard
-	if (stallGuardConfigured == false) {
-		stallGuardConfig();
-	}
-	stepper.moveTo(moveDist);
-	// if in fwd direction, reverse direction
-	if (directionFwd == true) {
-		changeDirection();
-	}
-	// while back position not set and forward position not set, run to -50000 or stall
-	// after stall, change direction and run to 50000 or stall
-	while (bwdPosition != 0 || fwdPosition <= 9999) {
-		stepper.run();
-	}
-	// if back and forward position set, move to middle position
-	if (bwdPosition == 0 && fwdPosition > 10000) {
-		bootFlag = false;
-		stepper.moveTo(fwdPosition / 2);
-		while (stepper.distanceToGo() != 0) {
-			stepper.run();
-		}
-	}
-}
-
-void dryRun() {
-  // wake stepper from sleep
-	if (digitalRead(EN_PIN) == HIGH) {
-    toggleStepper(1);
-  }
-
-  // move to start
-  while (stepper.currentPosition() > startPosition) {
-    stepper.setSpeed(-600);
-    stepper.moveTo(startPosition);
-    stepper.runSpeedToPosition();
-  }
-  displayPosition();
-
-  // step slow through procedure
-  while (stepper.currentPosition() < endPosition) {
-    stepper.setSpeed(100);
-    stepper.moveTo(endPosition);
-    stepper.runSpeedToPosition();
-  }
-  displayPosition();
-
-  // return to start
-  while (stepper.currentPosition() > startPosition) {
-    stepper.setSpeed(-600);
-    stepper.moveTo(startPosition);
-    stepper.runSpeedToPosition();
-  }
-
-  // stops motor instantly
-  stepper.setSpeed(0);
-  targetFlag = true;
-  movementProgress = false;
-  displayPosition();
-}
-
-bool stepperStep(int stepDirection, unsigned long stepperDelay) {
-  currentTime = millis();
-
-  // step after elapsed amount of time
-  if (currentTime - prevStepTime > stepperDelay) {
-    // wake stepper from sleep
-		if (digitalRead(EN_PIN) == HIGH) {
-	    toggleStepper(1);
-	  }
-
-    int stepVelocity = stepDirection * stepsPerMovement;
-
-    stepper.move(stepVelocity);
-		stepper.setSpeed(200*stepDirection);
-		while (stepper.distanceToGo() != 0) {
-	    stepper.runSpeed();
-			// stepDue = true;
-		}
-
-    prevStepTime = currentTime;
-		return true;
-  } else {
-			return false;
-	}
-}
-
-void joyStick() {
-  // wake stepper from sleep
-	if (stepperDisabled == true) {
-    toggleStepper(1);
-  }
-  // while joystick is operated
-  while (xStickPos >= xStickUpper || xStickPos <= xStickLower) {
-    // move either -1, 0, or 1 steps
-    stepper.move(map(xStickPos, 0, 1023, 1, -1));
-		// xStickMid = resting stable point of joystick§
-    joystickSpeed = map((xStickPos - xStickMid), -xStickMid, xStickMid, 2000, -2000);
-
-    stepper.setSpeed(joystickSpeed);
-    stepper.runSpeed();
-
-    // check stick position again
-    if (millis() % 50 == 0) {
-       xStickPos = analogRead(XSTICK_PIN);
-    }
-  }
-
-  // // check start/end position adjustment
-  if (editStartPosition == true && arrowsActive == true) {
-    if (prevStartPosition != startPosition) {
-      prevStartPosition = startPosition;
-    }
-    setAutoStackPositions(1, 0); //set start but not end position
-  }
-  if (editEndPosition == true && arrowsActive == true) {
-    if (prevEndPosition != endPosition) {
-      prevEndPosition = endPosition;
-    }
-    setAutoStackPositions(0, 1); //set end but not start position
-  }
-  if (activeScreen == 2 || activeScreen == 4) {
-    displayPosition();
-  }
-  // update prev position
-  prevStepperPosition = stepper.currentPosition();
-}
-
-void toggleJoystick() {
-  if ((currentTime - prevGenericTime) >= genericTouchDelay) {
-    joystickState = !joystickState;
-    prevGenericTime = millis();
-  }
-}
-
-// void toggleShutter() {
-//   if ((currentTime - prevGenericTime) >= genericTouchDelay) {
-//     shutterState = !shutterState;
-//     prevGenericTime = millis();
-//   }
-// }
-
-void resetAutoStack() {
-  if ((currentTime - prevGenericTime) >= genericTouchDelay) {
-    autoStackFlag = false;
-    goToStart = true;
-    joystickState = true;
-    pauseAutoStack = false;
-    // stepDue = true;
-    stepCount = 1;
-    movementProgress = false;
-    numMovements = 0;
-    startPosition = stepper.currentPosition();
-    endPosition = 0;
-    tft.drawBitmap(250, 190, reset40, 40, 40, GRAY);
-
-    prevGenericTime = millis();
-  }
-}
-
-// void setShutterDelay() {
-//
-//   if (shutterDelay < 1) {
-//     shutterDelay = 1;
-//   } else if (shutterDelay > 15) {
-//     shutterDelay = 15;
-//   }
-//
-//   if (prevDelay != shutterDelay) {
-//     // print new delay value
-//     tft.setCursor(155, 143);
-//     tft.setFont(&Arimo_Regular_20);
-//     tft.setTextColor(BLACK, BLACK);
-//     tft.println(prevDelay);
-//     tft.setCursor(155, 143);
-//     tft.setTextColor(WHITE, BLACK);
-//     tft.println(shutterDelay);
-//   }
-// }
-
-// bool triggerShutter() {
-// 	flashReady = flashStatus();
-// 	shutterTriggered = false;
-//   if (shutterState == true) {
-// 		unsigned long triggerTime = millis();
-//
-// 		// wait for flash to be ready
-// 		while (flashReady == false) {
-// 			flashReady = flashStatus();
-// 			delay(50);
-// 		}
-//
-// 		// trigger flash
-//     digitalWrite(SHUTTER_PIN, HIGH);
-//
-// 		// wait for flash to be triggered
-// 		while (shutterTriggered == false) {
-// 			delay(50);
-// 			// if signal missed or shot never taken, break
-// 			if (millis() - triggerTime > 10000) {
-// 				break;
-// 			}
-// 			// check if flash triggered
-// 			flashReady = flashStatus();
-// 			// if signal received, exit and reset shutter
-// 			if (flashReady == false) {
-// 				shutterTriggered = true;
-// 				delay(250); // pause for 250ms
-// 			}
-// 		}
-// 		recycleTime = (millis() - triggerTime);
-//
-// 		// reset shutter signal
-//     digitalWrite(SHUTTER_PIN, LOW);
-// 	}
-// 	return shutterTriggered;
-// }
-
-void setStepDistance() {
-
-  // constrain multiplier range
-  if (stepsPerMovement < 1) {
-    stepsPerMovement = 1;
-  } else if (stepsPerMovement > 80) {
-    stepsPerMovement = 40;
-  }
-
-  // stepper is 200 steps/rev, linear rail has 1mm pitch, 1:1 gear ratio
-  // 1 step = 1/200 = 0.005mm
-	// microsteps = 4
-	// 0.005 / 4 = 1.25μm
-  distancePerMovement = microstepDistance * stepsPerMovement;
-
-  if (prevDistance != distancePerMovement) {
-    // print new delay value
-    int16_t x1, y1;
-    uint16_t w, h;
-    tft.setFont(&Arimo_Bold_30);
-    tft.getTextBounds(String(prevDistance/1000, 4), 10, 60, &x1, &y1, &w, &h);
-    tft.fillRect(x1, y1, w, h, CUSTOM_BLUE);
-    tft.setCursor(10, 60);
-    tft.setTextColor(YELLOW, CUSTOM_BLUE);
-    tft.println(String(distancePerMovement/1000, 4));
-
-    // calculate number of steps required to cover range
-    numMovements = ceil((endPosition - startPosition)*1.00/stepsPerMovement);
-
-    if (activeScreen == 3) {
-      // update estimated time
-      estimateDuration(0); // don't force refresh
-      // update progress
-      updateProgress(0); // don't force refresh
-    }
-
-    prevNumMovements = numMovements;
-  }
-}
-
-void setAutoStackPositions(bool setStart, bool setEnd) {
-  if (setStart == 1) {
-    // lower limit
-    if (startPosition < 0) {
-      startPosition = 0;
-    }
-    // get new value
-    startPosition = stepper.currentPosition();
-    // print start point if changed
-    if (prevStartPosition != startPosition && activeScreen == 4) {
-      int16_t x1, y1;
-      uint16_t w, h;
-      tft.getTextBounds(String(prevStartPosition), 35, 65, &x1, &y1, &w, &h);
-      tft.fillRect(x1, y1, w, h, BLACK);
-      tft.setCursor(35, 65);
-      tft.setFont(&Arimo_Regular_16);
-      tft.setTextColor(WHITE, BLACK);
-      tft.println(startPosition);
-    }
-    // reset macroStepping
-    goToStart = true;
-  }
-
-  if (setEnd == 1) {
-    // lower limit
-    if (endPosition < 0) {
-      endPosition = 0;
-    }
-    // get new value
-    endPosition = stepper.currentPosition();
-    // print end point if changed
-    if (prevEndPosition != endPosition && activeScreen == 4) {
-      int16_t x1, y1;
-      uint16_t w, h;
-      tft.getTextBounds(String(prevEndPosition), 35, 145, &x1, &y1, &w, &h);
-      tft.fillRect(x1, y1, w, h, BLACK);
-      tft.setCursor(35, 145);
-      tft.setFont(&Arimo_Regular_16);
-      tft.setTextColor(WHITE, BLACK);
-      tft.println(endPosition);
-    }
-  }
-
-  // calculate number of steps required to cover range
-  numMovements = ceil((endPosition - startPosition)/stepsPerMovement);
-  prevNumMovements = numMovements;
-}
-
-void displayPosition() {
-  // update for new values
-  if (prevStepperPosition != stepper.currentPosition()) {
-    int16_t x1, y1;
-    uint16_t w, h;
-
-    if (activeScreen == 2) { // 2 = manual screen
-      // update rail position value
-      tft.setFont(&Arimo_Bold_30);
-      tft.getTextBounds(String(prevStepperPosition*0.0025, 4), 9, 220, &x1, &y1, &w, &h);
-      tft.fillRect(x1, y1, w, h, CUSTOM_BLUE);
-      tft.setCursor(9, 220);
-      tft.setTextColor(WHITE, CUSTOM_BLUE);
-      tft.println(String(stepper.currentPosition()*0.0025, 4));
-
-      // update movement count for manual screen
-      manualMovementCount++;
-
-      tft.setFont(&Arimo_Bold_30);
-      tft.getTextBounds(String(prevManualMovementCount), 45, 140, &x1, &y1, &w, &h);
-      tft.fillRect(x1, y1, w, h, CUSTOM_BLUE);
-      tft.setTextColor(WHITE);
-      tft.setCursor(45, 140);
-      tft.println(manualMovementCount);
-
-      prevManualMovementCount = manualMovementCount;
-    }
-
-    if (activeScreen == 4) { // 4 = autoConfigScreen
-      tft.setFont(&Arimo_Regular_16);
-      tft.getTextBounds(String(prevStepperPosition), 35, 225, &x1, &y1, &w, &h);
-      tft.fillRect(x1, y1, w, h, BLACK);
-      tft.setCursor(35, 225);
-      tft.setTextColor(WHITE, BLACK);
-      tft.println(stepper.currentPosition());
-    }
-
-    prevStepperPosition = stepper.currentPosition();
-  }
-}
-
-void estimateDuration(bool screenRefresh) {
-  float duration = numMovements * (shutterDelay + (recycleTime/1000));
-  float elapsed = movementProgress * (shutterDelay + (recycleTime/1000));
-  float remaining = duration - elapsed;
-  int minutes = floor(remaining / 60);
-  int seconds = int(remaining) % 60;
-  char timeMinutesSeconds[5];
-
-  int16_t x, y;
-  uint16_t w, h;
-  tft.setFont(&Arimo_Bold_30);
-  tft.setTextColor(WHITE);
-  sprintf_P(timeMinutesSeconds, PSTR("%02d:%02d"), minutes, seconds);
-
-  if ((minutes != prevMinutes || seconds != prevSeconds || screenRefresh == 1) && activeScreen == 3) {
-    tft.getTextBounds(String(prevTimeMinutesSeconds), 15, 140, &x, &y, &w, &h);
-    tft.fillRect(x, y, w, h, CUSTOM_BLUE);
-    tft.setCursor(15, 140);
-    tft.println(timeMinutesSeconds);
-
-    prevMinutes = minutes;
-    prevSeconds = seconds;
-    // assign new time to prev variable
-    sprintf_P(prevTimeMinutesSeconds, PSTR("%02d:%02d"), minutes, seconds);
-  }
-}
-
-void updateProgress(bool screenRefresh) {
-  char progressMovements[10];
-
-  int16_t x, y;
-  uint16_t w, h;
-  sprintf_P(progressMovements, PSTR("%02d/%02d"), movementProgress, numMovements);
-
-  if ((movementProgress != prevMovementProgress) || numMovements != prevNumMovements || screenRefresh == 1) {
-    tft.getTextBounds(String(prevProgressMovements), 20, 220, &x, &y, &w, &h);
-    tft.fillRect(x, y, w, h, CUSTOM_BLUE);
-    tft.setFont(&Arimo_Bold_24);
-    tft.setTextColor(WHITE);
-    tft.setCursor(20, 220);
-    tft.println(progressMovements);
-
-    prevNumMovements = numMovements;
-    prevMovementProgress = movementProgress;
-    // assign new time to prev variable
-    sprintf_P(prevProgressMovements, PSTR("%02d:%02d"), movementProgress, numMovements);
-  }
-}
-
-void toggleStepper(bool enable) {
-	if (enable == 1) {
-		delay(10); // give time for last step to complete
-		stepper.enableOutputs();
-		stepperDisabled = false;
-		delay(10); // breathing space
-	}
-
-	if (enable == 0) {
-		stepper.setSpeed(0);
-		stepper.move(0);
-		stepper.disableOutputs();
-		stepperDisabled = true;
-	}
-}
-
-// void stallGuardConfig() {
-//   stallGuardConfigured = true;
-//   //set TMC2130 config
-//   driver.push(); // reset registers
-//   driver.toff(3);
-//   driver.tbl(1);
-//   driver.hysteresis_start(8);
-//   driver.hysteresis_end(1);
-//   driver.diag1_stall(1);
-//   driver.diag1_active_high(1);
-//   driver.coolstep_min_speed(0x0009F); // 20bit max
-//   driver.semin(3);
-//   driver.semax(2);
-//   driver.sg_min(1); // if sg_result < sg_min*32, current increase
-//   driver.sg_max(3); // if sg_result >= (sg_min+sg_max+1)*32, current decrease
-//   driver.sedn(0b01);
-//   driver.sg_stall_value(0);
-//   driver.stealthChop(0);
-// }
-//
-// void silentStepConfig() {
-//   stallGuardConfigured = false;
-//   driver.push(); // reset registers
-//   driver.stealthChop(1);
-//   driver.stealth_autoscale(1);
-//   driver.stealth_gradient(0xF); // 1 to 15
-//   driver.interpolate(1);
-//   driver.push();
-//   driver.diag1_stall(0);
-//   driver.diag1_active_high(0);
-//   driver.coolstep_min_speed(0x0);
-//   driver.stealth_freq(0x0); // 0 or 1 for 16MHz
-// //    driver.chopper_mode(0);
-// //    driver.stealth_max_speed(0x0002F);
-// //    driver.double_edge_step(1);
-// //    driver.chopper_mode(1);
-// //    driver.sync_phases(1);
-// }
-
-// bool flashStatus() {
-//   int flashValue = 0;
-//
-//   flashValue = analogRead(FLASH_PIN);
-//   Serial.println(flashValue);
-//
-//   if (flashValue >= 280) {
-//     flashReady = true;
-//   }
-//   else {
-//     flashReady = false;
-//   }
-//
-//   return flashReady;
-// }
-
-void changeDirection() {
-  if (directionFwd == true) {
-    moveDist = -500000;
-    directionFwd = false;
-  }
-  else if (directionFwd == false) {
-    moveDist = 500000;
-    directionFwd = true;
-  }
-  stepper.moveTo(moveDist);
-}
-
-void stallDetection() {
-  Serial.print("DIAG: ");
-  Serial.println(digitalRead(DIAG_PIN));
-  if (bootFlag == true) {
-    toggleStepper(0);
-    stepper.setSpeed(0);
-    if (directionFwd == false) {
-      stepper.setCurrentPosition(0);
-      bwdPosition = stepper.currentPosition();
-    }
-    if (directionFwd == true) {
-      fwdPosition = stepper.currentPosition();
-    }
-    toggleStepper(1);
-    changeDirection();
   }
 }
